appSlice Definition createSlice Function:
createSlice is called with an object that defines a Redux slice.
name: This specifies the name of the slice, which is "app" in this case.
initialState: This defines the initial state of the slice. It includes several fields:
open: Initialized to true, likely used to control the state of a sidebar or modal.
video: Initialized as an empty array, presumably intended to hold video-related data.
category: Initialized to "All", possibly used to filter or categorize content.
searchSuggestion: Initialized as an empty array, likely intended to store search suggestions.