reducers Object:
This is an object containing Redux reducer functions.
toggleSidebar: A reducer function that toggles the open field between true and false.
setHomeVideo: A reducer function that sets the video field to the payload provided in the action.
setCategory: A reducer function that sets the category field to the payload provided in the action.
setSearchSuggestion: A reducer function that sets the searchSuggestion field to the payload provided in the action.