Imports The code starts with import statements:
React is imported from 'react'. This is necessary to use JSX syntax and React components.
ReactDOM is imported from 'react-dom/client'. It is used here for its createRoot method to render the React application.
App is imported from './App'. This presumably imports the main component of the application.
Provider is imported from 'react-redux'. It is a React component provided by React Redux that makes the Redux store available to the rest of the application.